See Description
There are a lot of changes here. Most notably, I updated the svgs to be read in from assets. This helps keep the relevant html portable and readable. It also helps me abstract away the issue with viewports and such into easier to understand measures (i.e. width and height). I also removed the width and height attributes from the svgs to be web-safe with safari. Before, spacing was an issue. I also added a reading time attribute to the post section.

Also started linear algebra blog and fixed minor typos in the previous blogs